organized the files even better. Still need to break out components into smaller pieces.
