+++ client/src/components/pages/Profile.js
@@ -1,9 +1,9 @@
 import React, {useEffect, useContext} from 'react'
-import TaskForm from './TaskForm.js'
-import Task from './Task'
-import { UserContext } from "../context/UserProvider.js"
-import { ProjectContext } from '../context/ProjectProvider'
-import ProjectList from './ProjectList'
+import TaskForm from '../Task/TaskForm.js'
+import Task from '../Task/Task'
+import { UserContext } from "../../context/UserProvider.js"
+import { ProjectContext } from '../../context/ProjectProvider'
+import ProjectList from '../projects/ProjectList'
 import Split from 'react-split'
 import {
     Box,
